requirements.txt
asgiref==3.5.2
Babel==2.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
django-money==3.0.0
django-rest-authtoken==2.1.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
py-moneyed==2.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
